requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
psycopg2-binary
drf-yasg
whitenoise